requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
idna==3.6
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
SQLAlchemy==2.0.25
alembic==1.13.1
python-jose==3.3.0
passlib==1.7.4
pydantic-settings==2.1.0
asyncpg==0.29.0
python-multipart==0.0.6
typer==0.9.0
httpx==0.26.0
bcrypt==4.1.2
SQLAlchemy-Utils==0.41.1
psycopg==3.1.17
psycopg-binary==3.1.17
psycopg2-binary==2.9.9
